requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-registration==3.1
django-tinymce==3.0.2
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.5.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pyuploadcare==2.6.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
